Create PS_Ping-FQDM-IP-NSLookup.ps1
<# This script was design to check if a server is up witha ping test, than to get it's FQDN,
and it's IP addressn than to get registered DNS name and IP through an NSLookup.

The information may seem redundant, but came about from a large and old server
environment where it was believed there were old DNS entrys and re-used IPs which
didn't reflect the current state of the network. #>