Create PS_Get_Remote_Event_Logs_Crit-Error-Warning.ps1
<# This script will get filtered system event logs for Critial, Error, and Warning alerts.
This script can take a while to complete, so give it 10 to 20 minuites if needed.
The script will output the file to an Excel file, with the server's name and today's date.
It will import an excel module and will create the path for the output if one is not created,
or you can list the path you want the output to got to. #>